Make sure we use _SpecialVector in FunctionAssigner - and add a test
