+++ tests/test_assign.py
@@ -27,6 +27,54 @@ def mesh_3D():
     return dolfinx.mesh.create_unit_cube(MPI.COMM_WORLD, 50, 50, 50)
 
 
+def test_multiple_assign_adjoint_accumulation(mesh_1D):
+    """
+    Test that assigning a Function multiple times and computing the adjoint
+    successfully accumulates the adjoints. This tests the IAddableVector patch.
+    """
+    pyadjoint.get_working_tape().clear_tape()
+    mesh = mesh_1D
+
+    V = dolfinx.fem.functionspace(mesh, ("Lagrange", 1))
+
+    # 1. Create a control function
+    f = Function(V, name="f_control")
+    f.interpolate(lambda x: 2.0 * x[0])
+
+    # 2. Fork the computational graph by assigning f to two different functions
+    u1 = Function(V, name="u1")
+    assign(f, u1)
+
+    u2 = Function(V, name="u2")
+    assign(f, u2)
+
+    # 3. Assemble a functional depending on both branches
+    J_form = (u1**2 + u2**2) * ufl.dx(domain=mesh)
+    J = assemble_scalar(J_form)
+
+    control = pyadjoint.Control(f)
+    Jhat = pyadjoint.ReducedFunctional(J, control)
+
+    # 4. Compute the derivative
+    # Before the patch, this would crash because pyadjoint attempts to sum
+    # the adjoint contributions from u1 and u2 using `+=` on dolfinx.la.Vector.
+    # After the patch, IAddableVector handles this gracefully.
+    grad = Jhat.derivative()
+    assert grad is not None
+
+    # 5. Verify the math with a Taylor test
+    df = Function(V)
+    df.interpolate(lambda x: np.sin(x[0]))
+
+    # Without gradient
+    min_rate_0 = pyadjoint.taylor_test(Jhat, f, df, dJdm=0)
+    assert np.isclose(min_rate_0, 1.0, rtol=1e-2, atol=1e-2)
+
+    # With gradient
+    min_rate_1 = pyadjoint.taylor_test(Jhat, f, df)
+    assert np.isclose(min_rate_1, 2.0, rtol=1e-2, atol=1e-2)
+
+
 @pytest.mark.parametrize("constant", [np.float64(0.2), float(-0.13), int(3)])
 @pytest.mark.parametrize("mesh_var_name", ["mesh_1D", "mesh_2D", "mesh_3D"])
 def test_assign_constant(mesh_var_name: str, request, constant: typing.Union[float, int, np.floating]):
